TG-150: fix logout logic
