+++ src/components/Shared/Header/index.js
@@ -64,7 +64,8 @@ const Header = () => {
             className={styles.userButton}
             onClick={() => {
               dispatch(logout());
-              history.push('/login');
+              history.push('/home');
+              window.location.reload();
             }}
           >
             <p className={styles.text}>Log out</p>
