+++ src/redux/auth/thunks.js
@@ -51,6 +51,5 @@ export const logout = () => {
     firebase.auth().signOut();
     sessionStorage.clear();
     dispatch(logOut());
-    window.location.reload();
   };
 };
